# docs/openapi/pn-normalizzatore-webhook-v1.yaml
openapi: 3.0.3
info:
  title: Normalizzatore ingress API.
  description: >-
    Normalizzazione degli indirizzi. L'API espone dei servizi sul path "/normalizzatore-ingress/v1/".
  version: v1.0
servers:
  - url: >-
      https://api.externalchannel.pagopa.local
    description: Server url
tags:
  - name: Normalizzatore
    description: API esposte verso il normalizzatore

paths:
  /normalizzatore-ingress/v1/callback:
    post:
      summary: Richiesta di callback di avvenuta normalizzazione della lista di address in ingresso
      description: |-
        Operazione che avvisa pn-address-manager dell'avvenuta normalizzazione di una lista di address.
      tags:
        - Normalizzatore
      parameters:
        - $ref: '#/components/parameters/xPagopaAddrManCxId'
        - $ref: '#/components/parameters/xPagopaAddrManApiKey'
      operationId: callbackNormalizedAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackRequestData'
      responses:
        '200':
          description: Risposta di successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponseData'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/CallbackResponseData'
        '500':
          description: InternalServerError
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/CallbackResponseData'

  /normalizzatore-ingress/v1/attachment-preload:
    put:
      summary: Richiesta di pre-caricamento dei documenti da comunicare a PN
      description: |-
        Operazione che richiede a Piattaforma Notifica le informazioni e le autorizzazioni necessarie 
        a precaricare uno o più file.
      tags:
        - Normalizzatore
      parameters:
        - $ref: '#/components/parameters/xPagopaAddrManCxId'
        - $ref: '#/components/parameters/xPagopaAddrManApiKey'
      operationId: presignedUploadRequest
      requestBody:
        content:
          application/json:
            schema:
              title: PreLoadRequestData
              type: object
              required:
                - preloads
              properties:
                preloads:
                  type: array
                  items:
                    $ref: "#/components/schemas/PreLoadRequest"
                  minItems: 1
                  maxItems: 15
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                title: PreLoadResponseData
                type: object
                required:
                  - preloads
                properties:
                  preloads:
                    type: array
                    items:
                      $ref: "#/components/schemas/PreLoadResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResultCodeResponse'
              examples:
                syntaxError:
                  summary: 'Errore sintattico'
                  value: { resultCode: '400.01', resultDescription: 'Syntax Error', errorList: [ 'field requestId is required' ], clientResponseTimeStamp: '2019-08-24T14:15:22Z' }
                semanticError:
                  summary: 'Errore semantico'
                  value: { resultCode: '400.02', resultDescription: 'Semantic Error', errorList: [ 'unrecognized product Type' ], clientResponseTimeStamp: '2019-08-24T14:15:22Z' }

  
  /normalizzatore-ingress/v1/get-attachment/{fileKey}:
    get:
      summary: ottieni metadati e contenuto di un file.
      description: |-
        Con questa API è possibile ottenere le informazioni necessarie al download del contenuto del file;
        tali informazioni sono contenute nel campo _download_. <br/>
          - _url_: se il contenuto del file è immediatamente disponibile questo campo viene 
            valorizzato con un URL da invocare in GET per ottenere il contenuto del file. <br/>
          - _retryAfter_: valorizzato se il contenuto del file è stato archiviato in glacier. 
            Indica una stima del tempo necessario per recuperare il file. <br/>
        
        Se il file non è pronto al momento dell'invocazione il campo _download.url_ della risposta
        verrà valorizzato ```null```; in un secondo momento, quando il contenuto del file sarà 
        pronto verrà comunicato un evento al client che ha invocato la chiamata. Molteplici 
        invocazioni di questa API generano un solo evento per client. <br/>
        Il file rimane hot per ```stayHotTime````.
      parameters:
        - $ref: '#/components/parameters/fileKey'
        - $ref: '#/components/parameters/xPagopaAddrManCxId'
        - $ref: '#/components/parameters/xPagopaAddrManApiKey'
      tags:
        - Normalizzatore
      operationId: getFile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownloadResponse'
              examples:
                hotFile:
                  summary: File pronto per il download
                  value:
                    key: 'random/path/of/the/file'
                    versionId: '3Z9SdhZ50PBeIj617KEMrztNKDMJj8FZ'
                    contentType: 'application/pdf'
                    contentLength: 30438
                    checksum: '91375e9e5a9510087606894437a6a382fa5bc74950f932e2b85a788303cf5ba0'
                    retentionUntil: '2032-04-12T12:32:04.000Z'
                    documentType: PN_LEGALFACT
                    documentStatus: SAVED
                    download:
                      url: 'https://presignedurldemo.s3.eu-west-2.amazonaws.com/...'
                coldFile:
                  summary: File su glacier.
                  description: Il file è stato archiviato. Verrà recuperato in 24 ore.
                  value:
                    key: 'random/path/of/the/file'
                    versionId: '3Z9SdhZ50PBeIj617KEMrztNKDMJj8FZ'
                    contentType: 'application/pdf'
                    contentLength: 30438
                    checksum: '91375e9e5a9510087606894437a6a382fa5bc74950f932e2b85a788303cf5ba0'
                    retentionUntil: '2032-04-12T12:32:04.000Z'
                    documentType: PN_LEGALFACT
                    documentStatus: SAVED
                    download:
                      retryAfter: 86400
        '404':
          description: >-
            Non trovato. Restituito se: <br/>
            - il parametro clientId è diverso dal client autenticato; <br/>
            - se il file identificato da fileKey non è ancora stato salvato; <br/>

components: 

  parameters: 
    fileKey: 
      name: fileKey
      description: La chiave definita per il file di cui si sta richiedendo il download
      in: path
      required: true
      schema:
        type: string
    xPagopaAddrManCxId:
      $ref: 'parameters-pn-address.yaml#/components/parameters/xPagopaAddrManCxId'
    xPagopaAddrManApiKey:
      $ref: 'parameters-pn-address.yaml#/components/parameters/xPagopaAddrManApiKey'
  
  schemas: 

    ##########################################################################
    ###                 PRECARICAMENTO FILE SU SAFESTORAGE                 ###
    ##########################################################################

    PreLoadRequest:
      title: Richiesta di precaricamento di un File
      type: object
      properties:
        preloadIdx:
          title: Id della richiesta di precaricamento di un file
          description: >-
            Identificativo univoco all'interno della request HTTP, serve per correlare la risposta. 
          type: string
          minLength: 1
          maxLength: 250
        contentType:
          title: MIME type del file che verrà caricato
          description: >-
            Il MIME type dell'allegato che dovrà essere caricato. 
          type: string
          minLength: 3
          maxLength: 127
          example: application/pdf
        sha256:
          minLength: 40
          maxLength: 50
          title: checksum sha256 del file che verrà caricato
          description: >-
            checksum sha256, codificato in base 64, del contenuto binario del file che verrà
            caricato
          type: string
          example: jezIVxlG1M1woCSUngM6KipUN3/p8cG5RMIPnuEanlE=

    PreLoadResponse:
      title: Informazioni per il caricamento file
      description: >-
        Per ogni richiesta che è stata fatta viene fornito un presigned URL e le 
        informazioni per usarlo.
      type: object
      properties:
        preloadIdx:
          description: per correlazione con la richiesta
          type: string
          minLength: 1
          maxLength: 250
        secret:
          description: >-
            Token aggiuntivo per far si che sia necessario intercettare anche gli 
            header e non solo l'URL di upload del contenuto del documento.
          example: AZ23RF12
          minLength: 1
          maxLength: 250
          type: string
        httpMethod:
          description: >-
            Indica se per l'upload del contenuto file bisogna utilizzare il metodo PUT o POST
          type: string
          example: PUT
          enum:
            - POST
            - PUT
        url:
          description: >-
            URL a cui effettuare l'upload del contenuto del documento.
          type: string
          example: 'https://preloadpn.aws.amazon.......'
          minLength: 5
          maxLength: 2000
        key:
          description: >-
            la chiave restituita sarà globalmente unica e andrà utilizzata nella richiesta 
            di notifica.
          example: 'PN_NOTIFICATION_ATTACHMENTS-0001-301W-B9CB-9U72-WIKD'
          type: string
          minLength: 30
          maxLength: 250

    FileDownloadResponse:
      description: Informazioni necessarie al download del contenuto di un file
      allOf:
        - $ref: '#/components/schemas/FileCreatedDto'
        - type: object
          properties:
            download: 
              $ref: '#/components/schemas/FileDownloadInfo'
  
    FileCreatedDto:
      description: >-
        Evento generato quando il contenuto di un file è stato caricato, elaborato da 
        _SafeStorage_, ed è pronto all'uso.
      type: object 
      required:
        - key
        - versionId
        - contentType
        - contentLength
        - checksum
        - documentType
        - documentStatus
        - retentionUntil
      properties:  
        key:
          type: string
          description: la chiave da utilizzare per il recupero del file
          example: '8F7E/9A3B/1234/AB87'
        versionId:
          type: string
          description: __discutibile__. La versione del file su S3
          example: '3Z9SdhZ50PBeIj617KEMrztNKDMJj8FZ'
        documentType:
          type: string
          description: Il tipo del documento
          example: PN_NOTIFICATION_ATTACHMENTS
        documentStatus:
          type: string
          description: Lo stato del documento
          example: ATTACHED
        contentType:
          type: string
          example: 'application/pdf'
        contentLength:
          type: number
          example: 54092
          description: dmensione, in byte, del contenuto.
        checksum:
          type: string
          description: >- 
            SHA256 o MD5 del contenuto del file. Il tipo di checksum cambia a seconda del 
            documentType.
        retentionUntil:
          type: string
          format: date-time
          description: >-
            data e ora fino a cui il documento sarà garantito all'interno di _SafeStorage_. <br/>
            Può essere ritardata nel tempo, mai anticipata. Formattata con timezone secondo l'rfc 3339
                            
    FileDownloadInfo:
      type: object
      properties: 
        url:
          type: string
          description: >-
            URL preautorizzato a cui effettuare una richiesta GET per ottenere il 
            contenuto del documento. Presente solo se il documento è pronto per il download.
        retryAfter:
          type: number
          description: >-
            Stima del numero di secondi da aspettare prima che il contenuto del 
            documento sia scaricabile.


    CallbackRequestData:
      type: object
      properties:
        fileKeyInput:
          type: string
        fileKeyOutput:
          type: string

    CallbackResponseData:
      type: object
      properties:
        response:
          type: string
          enum:
            - OK
            - KO
    ActvatePostelRequest:
      type: object
      properties:
        fileKey:
          type: string

    ActvatePostelResponse:
      type: object
      properties:
        response:
          type: string
          enum:
            - OK
            - KO

    OperationResultCodeResponse:
      required:
        - resultCode
        - resultDescription  
        - clientResponseTimeStamp 
      type: object
      properties:
        resultCode:
          type: string
          minLength: 3
          maxLength: 10
          example: '400.00'
          description: >-
            Codice a 5 cifre separate da punto che indica l’esito della richiesta.  LE prime tre sono ripetizione del codice HTTP<br/>
            Per convenzione, la sequenza 2000 indicherà OK. <br />
            - '200.00' OK <br />
            - '400.01' Errore di validazione sintattica del messaggio <br />
            - '400.02' Errore di validazione regole semantiche <br />
            - '404.00' requestId mai ricevuto <br />
            - '409.00' requestId già utilizzato <br />
            - '500.xy' Errori interni al server <br />
            - .... Altri errori previsti dalla specifica implementazione <br />
        resultDescription:
          type: string
          minLength: 5
          maxLength: 250
          example: 'Request validation errors'
          description: >-
            Descrizione dell’Esito dell’operazione, sempre valorizzato, anche in caso di esito positivo. <br />
            In caso di errore dovrà contenere info utili al chiamante per indirizzare le azioni correttive necessarie. <br />
            In nessun caso conterrà dettagli tecnici e/o stack trace. <br />
        errorList: 
          type: array
          minItems: 0
          maxItems: 50
          items:
            type: string
            minLength: 3
            maxLength: 250
            example: 'unsupported printType value'
          description: >-
            Elenco degli errori di validazione, se presenti.
        clientResponseTimeStamp: 
          type: string
          format: date-time
          description: >-
            Timestamp della response in UTC  